generacize backend config
